blocker field narrower, for block-build-follow
If the blocker field had been 14 instead of 12 hi-res pixels wide into
each direction, there was a possibility for building with lix A,
blocking her with lix B on the staircase, A continuing to build, but the
crowd unable to follow. Now, with the narrower blocker field, the crowd
can always follow.