Remove unused cross tiles
Only the rusty cross is used, we'll keep that. It's the 4-way
intersection piece. Remove the steel 4-way intersection.

Remove the cross intersections, both steel and rust. You can create
those manually by overlapping two pipes. It's rare that you want it, and
it's okay then to require manual overlap of two pieces.