Fix 3-digit hatch count at 1920x1080
The bug was:

During play, you have the skillbar at the bottom, and the infobar above
the skillbar. In the infobar, there is a hatch icon and a 3-digit count
for number of lix in hatch. At 1920x1080 screen resolution, number of
lix in hatch abbreviated its third digit to a dot ".", e.g., 123 showed
as "12." instead of as "123".

Fix:

Widen the text field for number of skills in hatch. It overlaps with
the total-lix icon, but that icon is long and vertical, with
transparency at both left/right sides. You can't see overlap.