Show error box for resources not found
Show error boxes only in interactive mode.

Show error boxes even when we can't log. For example, we can't log when
we run Lix outside its directory tree. We only want to abort and tell
the player to run us from inside the tree.

Only on Windows, we ever show error boxes. Linux and other systems print
the exception to the command line and, if possible, log the exception.